requirements.txt
numpy==1.26.1
matplotlib==3.8.0
Pillow==10.0.1
diffusers
datasets==2.14.5
tqdm==4.66.1
torch==1.13.1+cu116
torchvision==0.14.1+cu116
tensorboard
rich
pytest
setuptools
omegaconf
